+++ tests/otree/settings.py
@@ -17,8 +17,9 @@ SESSION_FIELDS = []
 # for example: de, fr, ja, ko, zh-hans
 LANGUAGE_CODE = 'en'
 
+# Only relevant when OTREE_AUTH_LEVEL is set to "DEMO" or "STUDY"
 ADMIN_USERNAME = 'admin'
 ADMIN_PASSWORD = environ.get('OTREE_ADMIN_PASSWORD')
-SECRET_KEY = environ.get('OTREE_REST_KEY')
+SECRET_KEY = '12345'
 
 DEMO_PAGE_INTRO_HTML = """ """