+++ tests/test_b_otree.py
@@ -11,32 +11,27 @@ from tests.utils import *
 def test_botex_env_exists_in_project_root():
     assert os.path.exists("botex.env")
 
-@pytest.mark.dependency(name="otree_admin", scope='session', depends=["botex_env"])
-def test_botex_env_has_otree_admin_password():
+@pytest.mark.dependency(
+    name="otree_project_path", scope='session', depends=["botex_env"]
+)
+def test_botex_env_has_otree_project_path():
     botex.load_botex_env()
-    assert os.getenv("OTREE_ADMIN_PASSWORD")
+    assert os.getenv("OTREE_PROJECT_PATH")
 
-@pytest.mark.dependency(name="otree_rest_key", scope='session', depends=["botex_env"])
-def test_secret_has_otree_rest_key():
-    with open("botex.env") as f:
-        for line in f:
-            if "OTREE_REST_KEY" in line:
-                assert line.split("=")[1]
-
-@pytest.mark.dependency(name="otree_starts")
+@pytest.mark.dependency(name="otree_starts", depends=["otree_project_path"])
 def test_does_otree_start():
-    otree_proc = start_otree()
-    time.sleep(OTREE_STARTUP_WAIT)
+    botex.load_botex_env()
+    otree_proc = botex.start_otree_server()
     otree_running = otree_proc.poll() is None
-    if otree_running: stop_otree(otree_proc)
+    assert otree_running
+    if otree_running: botex.stop_otree_server(otree_proc)
 
 @pytest.mark.dependency(name="botex_session", scope='session', depends=["otree_starts"])
 def test_can_otree_session_be_initialized():
     botex.load_botex_env()
-    otree_proc = start_otree()
-    time.sleep(OTREE_STARTUP_WAIT)
+    otree_proc = botex.start_otree_server()
     botex_session = init_otree_test_session()
-    stop_otree(otree_proc)
+    botex.stop_otree_server(otree_proc)
     delete_botex_db()
     assert len(botex_session) == 5
     assert isinstance(botex_session["session_id"], str)
@@ -55,7 +50,7 @@ def test_can_otree_session_be_initialized():
 )
 def test_session_is_recorded_in_botex_db():
     delete_botex_db()
-    otree_proc = start_otree()
+    otree_proc = botex.start_otree_server()
     botex_session = init_otree_test_session()
     participants = botex.read_participants_from_botex_db(botex_db="tests/botex.db")
     assert isinstance(participants, list)
@@ -68,6 +63,10 @@ def test_session_is_recorded_in_botex_db():
     assert isinstance(p1["url"], str)
     assert p1["time_in"] == None
     assert p1["time_out"] == None
-    stop_otree(otree_proc)
+    botex.stop_otree_server(otree_proc)
     delete_botex_db()
+
+
+if __name__ == "__main__":
+    test_does_otree_start()